Decompile the Chair
Chair entity and helper function.

These required additional splits for rodata alignment.
