+++ config/splat.us.stno1.yaml
@@ -61,6 +61,8 @@ segments:
       - [0x34670, .rodata, e_mouse]
       - [0x34670, .rodata, e_red_door]
       - [0x34688, .rodata, unk_35E20]
+      - [0x346BC, .rodata, e_chair]
+      - [0x346D0, .rodata, unk_381E8]
       - [0x3479C, .rodata, unk_3B984]
       - [0x347BC, .rodata, unk_3CB34]
       - [0x347D8, .rodata, unk_3D8AC]
@@ -85,6 +87,8 @@ segments:
       - [0x34E6C, c, e_red_door]
       - [0x35C64, c, st_debug]
       - [0x35E20, c, unk_35E20]
+      - [0x38008, c, e_chair]
+      - [0x381E8, c, unk_381E8]
       - [0x3B984, c, unk_3B984]
       - [0x3CB34, c, unk_3CB34]
       - [0x3D8AC, c, unk_3D8AC]
